Modified UserController

+++ src/main/java/org/teapot/backend/controller/UserController.java
@@ -1,8 +1,10 @@
 package org.teapot.backend.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.teapot.backend.controller.exception.BadRequestException;
 import org.teapot.backend.controller.exception.ResourceNotFoundException;
 import org.teapot.backend.model.User;
 import org.teapot.backend.service.abstr.UserService;
@@ -20,8 +22,7 @@ public class UserController {
 
     @GetMapping
     public List<User> getUsers(@RequestParam(required = false) Integer offset,
-                               @RequestParam(required = false) Integer count,
-                               HttpServletResponse response) {
+                               @RequestParam(required = false) Integer count) {
         List<User> users;
         if ((offset != null) && (count != null)) {
             users = userService.getList(offset, count);
@@ -32,28 +33,32 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public User getUser(@PathVariable Long id, HttpServletResponse response) {
+    public User getUser(@PathVariable Long id) {
         User user = userService.getById(id);
         if (user == null) {
             throw new ResourceNotFoundException();
         }
-        response.setHeader("Location", "/api/users/" + id);
         return user;
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void putUser(@PathVariable Long id,
-                        @RequestBody User user,
-                        HttpServletResponse response) {
+    public void updateUser(@PathVariable Long id,
+                        @RequestBody User user) {
+        System.out.println();
+        if (userService.getById(id) == null) {
+            throw new ResourceNotFoundException();
+        }
         user.setId(id);
         userService.update(user);
-        response.setHeader("Location", "/api/users/" + id);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteUser(@PathVariable Long id) {
+        if (userService.getById(id) == null) {
+            throw new ResourceNotFoundException();
+        }
         userService.delete(id);
     }
 
@@ -61,7 +66,11 @@ public class UserController {
     @ResponseStatus(HttpStatus.CREATED)
     public User registerUser(@RequestBody User user,
                              HttpServletResponse response) {
-        userService.register(user);
+        try {
+            userService.register(user);
+        } catch (DataIntegrityViolationException e) {
+            throw new BadRequestException();
+        }
         response.setHeader("Location", "/api/users/" + user.getId());
         return user;
     }
